Update patch set 1
No score.

Error handling in init is, as you note, a bit inconsistant.

I'd say, return -1 on failure for new commands.  

We should update old commands to do the same, and update the command
dispatcher to at the very least log failed commands --  but these should
be addressed in separate patches.

Regarding updating patches, there is a repo/gerrit feature in the works
to push a new patch to a change, and to allow reviewers to diff between
the original and updated versions (to more easily follow what changed). 
It's not ready yet, but coming soon.

Regarding this specific change -- do you need chdir/chroot or would just
providing pivotroot be more useful if the goal is to pivot to root on
another device?

Patch-set: 1