Update patch set 1
No score.

Sorry about the < and ( cases, I will fix those in a new patch.

However, the issues involving the fallthrough on the mount confuse me...
mtd@, the other similar structure that is only five lines above my added
loop@, does exactly that: if mtd_name_to_number fails it returns -1. The
check on this end of the code is for >= 0 and it does its work if and
only if it succeeds, otherwise it falls through the code and attempts to
mount "mtd@whatever" (which fails). That's exactly the paradigm I
followed in writing loop@ (as stupid as it seemed) as it seems more
correct that loop@ does the same thing, unless mtd@ is also changed, as
the "style" of the code is /not/ to check for < 0 and return an error as
you claim :(.

However, handling the case of unmounting the loopback device will make
that more awkward: to keep the current style in the code my only thought
is to keep the loop variable in use until after the call to mount (with
a default value of -1 for the cases where it isn't used by loop@) and
then check if it is >= 0 and, if so, close it and possibly (if mount
failed) disassociate it.

It therefore makes more sense, probably, to refactor the entire function
in a single go, and fix the mtd@ case as well. I think I'll go ahead and
do that and hope it isn't thrown away as "too large of a patch". :(

I have done this and uploaded the new patch to here:
http://review.source.android.com/4251

Patch-set: 1