Update patch set 1
No score.

I went through and redid the entire function in order for the style to
be consistent with the code that was already there (which, if it had
been committed now, supposedly wouldn't have gotten in as it violated
the style). I submitted this as a different patch, and am now trying to
look into better options for getting rid of this one than "delete"ing
it. I am therefore making a semi-useful comment so that I can click
"Abandoned" and see what it does. I apologize if the "semi-" on this
comment's usefulness bothers someone. :( Repo and Gerrit need better
documentation. Ooo...  I can uncheck "Send mail" maybe.

Patch-set: 1