Create patch set 1
Patch-set: 1
Subject: Small clean up and fix to the problem of having using insmod in the init.rc of drivers that expect to use udev to load driver FW.  Without this fix the driver will block on the FW load waiting for the init program to respond to the netlink message, while its waiting on the insmod call to complete. Loading driver firmware as a result of using insmod in the init.rc is a multi-threaded operation.
Commit: da0ab7f99cd130bebd652063188e5224425f3618