Update patch set 1
No score.

Re busybox:
At least on rc30, iptables is not shipped on the device. (Admittedly,
neither is busybox, but it is on a very popular alternative image.) It
is simple enough to change the path anywhere the iptables binary is on a
particular system.

As far as hardcoded IP for the phone (acting as a server/router), what
would you suggest instead? It needs to be put on the interface and
present in udhcpd.conf and it must be in the block of private addresses.

Patch-set: 1