Update patch set 1
No score.

No score.

(Still trying, since the site is giving 500 errors. :( If this works,
then 'reply' is seriously broken.)

On 2008/11/29 09:18:03, "Alexey Tarasov" <tarasov wrote:
> Anyway, I think it's more useful to make something like:
> ---
> ---
>
> where INT_IP/EXT_IP may be got via parsing of ifconfig otput or
> udhcpd.leases file.

You have it backwards here - the phone is a dhcp -server-, and so there
is no address to be parsed until we get running. ESPECIALLY in the
init.rc, which is - as far as I understand it - a fairly bizarre little
scripting language of it's own.

> Main point in usage abstract ${IPTABLES} instead of explicit command
is
> that in such case you may choose exact command even automatically by
> checking existance of busybox, thus script will be more reusable.

This is an embedded device - indirection for indirection's sake is
frowned upon. We're not going to have a bunch of endusers installing
different core utilities in different places and generally making life
interesting. The bins are set by the system, and when that happens the
commands can be updated (once) and done.

> Also, As far as I know you may just specify interfaces, instead of
IPs:
> ---
> ---

Not in postrouting.
# iptables -t nat -A POSTROUTING -i eth0 -o eth1 -j ACCEPT
iptables v1.3.8: Can't use -i with POSTROUTING

Patch-set: 1