Update patch set 1
I would prefer that you didn't submit this.

I have to agree with Brian here. Init is important enough to try to keep
it simple, and multithreadedness is never simple.

Since it should be easy to fork/exec/waitpid a separate multithreaded
process if really necessary (there's no memory pressure at that point,
though of course the cost of creating a new process is unfortunate) I
think that process 1 itself could stay single-threaded.

There might not even be a need to actually exec anything, putting the
relevant code in init and just forking the multithreaded part would do
the trick.

Patch-set: 1