init: add "androidboot.bootmedia" kernel command line argument
support boot specific init.<bootmedia>.rc for boot from diverse boot media.
this argument can be "sdcard" or "harddisk" or other else and if this
argument is not handed over the bootloader, "sdcard" is used as the default
value. The init process loads init.<bootmedia>.rc file first according to
"androidboot.bootmedia=" and it loads init.rc later.