Neon implementation of t32cb16blend function
16-pixel per pass SIMD implementation, little endian only. Based on arm
implementation, guards for selection only if TARGET_ARCH_VERSION is armv7-a,
__ARM_NEON__ is defined (this is supplied by GCC when -mfpu=neon is passed to
the compiler) and the system is little endian.

Tested on ZoomII using soft-keyboard and web browser applications.

It's possible for a system to support ARMv7-a, but not have Neon available -
this is the reason for 'optimistically' including the file at the Makefile
level but using __ARM_NEON__ to actually select the code for compilation.

Uses ARCH_ARM rather than TARGET_ARCH to select NEON features
at build on a per-capability rather than per-arch