Fix issue of procfs and usb for FreeBSD.
The format of linkmap exported by procfs of FreeBSD is different from
what is exported by procfs of Linux.  This changes fix the issue for
FreeBSD.

FreeBSD's implementation of FreeBSD is also different from Linux.
This changes provide a dummy implementation to make the system can be
built.  But, USB is still not working on FreeBSD.