Update patch set 1
Patch Set 1: No score; Looks good to me, but someone else must approve

I'm inclined to let those in. They're cleanly separated in a way that doesn't put any risk on the rest of the system. Sure, we (Google/AOSP) won't be able to support them, but I assume that's an understood risk.

Patch-set: 1
Label: Code-Review=+1