Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

This patch is either not needed, or it will not work. (It may compile, but the result will fail to run correctly.)

The reason is that the code in acc.c has many assumptions that sizeof(void*) == 4. For example, acc generates 32 bit x86 code and it assumes that it can store pointers in the 32-bit address portion of the x86 code. Obviously there is no way that a 64-bit pointer can fit in 32 bits. So code like this:

* (intptr_t*) x = foo

will do the wrong thing in acc if you run in a context where sizeof(intptr_t*) == 8 . (It will write 8 bytes instead of 4 bytes.)

Because of dependencies like this, acc.c only works in 32-bit contexts. And if you are in a 32-bit context you don't need to use intptr_t, you can use int.

You don't explain why you want to submit the patch in the first place, but I gather it's because you're getting some sort of compiler error when you compile under Snow Leopard.

I don't have Snow Leopard handy to reproduce the bug, but I think it might be possible to work around the problem by modifying the acc make file to make sure that the host acc tool is compiled as a 32-bit executable, rather than 64-bit executable.

(I'm assuming there's some way to do that with Snow Leopard.)

Patch-set: 1
Reviewer: Gerrit User 1003182 <1003182@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1