Update patch set 1
Patch Set 1: No score

The one adding the symlink stuff was not by me (look at
the author, which Gerrit doesn’t show, not the committer).

However, I think this is only a temporary problem on my
side which I only have to track down… if so I’ll take
back the comment-out commit, if not, both of them. But
I preferred to publish the entire history first thing,
but in a usable state.

Patch-set: 1
Reviewer: Gerrit User 1002670 <1002670@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0