Fix adb handling of devices with no serial number. See bug #4365.
If there is no serial number it will use the usb device name (/dev/usb/...).
This obviously isn't ideal, but it's better than before, and stops eclipse
crashing and burning.

Note also that I've only fixed it for Linux. I added a commented-out
possible solution for OSX but I don't have access to OSX to test it.
Added TODO notes to fix it on OSX and Windows. Someone else will have to do that though.

Possible reasons why not to merge, and responses:

1. We won't merge it until you fix OSX and Windows.
  - Ok! Buy me a Mac and a copy of Windows and I'd be happy to. Until then
    this is better than the current state and in no way worse. You might say
    it is a Pareto Patch. :-)
2. Change the comments.
  - There's quite a lot of overhead involved in contributing code to Android.
    If it were as simple as changing the test and clicking 'Save' then sure,
    otherwise it is much easier for people at Google to do this.
3. Looks good, we'll merge it.
  - Yeay! I shall submit more patches in future! :-)

Change-Id: I80d9a5941466c025092a0971582a4659f754f4f4