Use the external 'screencap' command to grab the screen
This series of patches changes the mechanism used by the DDMS debugger
to ensure that the screenshot tool always captures an accurate, atomic
image of the screen.

When the DDMS debugger requests a screenshot, try to grab it using an
external tool which knows how to talk to the SurfaceFlinger and fetch
an accurate screen image.

If this fails for any reason, fall back on reading directly from the
frame buffer.

Change-Id: Iae35ea77eb889da6a6867f21fe5701db3cbb3489