vold: scan only devices on same media when searching for partitions
Currently vold checks whether a partition has the same major as
the disk when matching partitions to block devices. This is not
only inefficient, but it also doesn't work for drivers like MMC
which use the same major for all device instances. Instead, just
search devices which have been associated with the same media as
the disk to determine whether vold knows about all partitions on
a disk.

Change-Id: I1a8c9ca56fbf50080bc176da6dbc38afd4f576b9