Update patch set 3
Patch Set 3: No score; I would prefer that you didn't submit this

The change merges and builds, and the emulator starts. It almost merges into Google's master tree as well, but I haven't investigated further.

I'm worried that very many things could be implicitly depending on the exact behavior of the existing shell, in a way that isn't documented anywhere (and possibly that isn't visible to the open-source world), and that could break in a way that can't really be tested ahead of time for not knowing what they are.

Patch-set: 3
Label: Code-Review=-1
Label: Verified=0