Add mksh-current from today’s CVS as system/core/mksh module.
This is disabled by default and will need $(TARGET_SHELL) to
be set to mksh to build. Verbose change history is available
at Gerrit; in contrast to patch set 3 of change 14171, the
ability to write to /dev/urandom via $RANDOM and the printf
builtin have been deactivated, and all unused files have been
removed from the source tree (this also simplifies the NOTICE
file), and links to the online manpage have been added.

Change-Id: I512aac5ba99768af5154f4889b805b88f290a8aa