Add mksh-current from CVS 2010/06/03 as system/core/mksh module.
This is disabled by default and will need $(TARGET_SHELL) to
be set to mksh to build. Verbose change history is available
at Gerrit; in contrast to patch set 3 of change 14171, the
ability to write to /dev/urandom via $RANDOM and the printf
builtin have been deactivated, and all unused files have been
removed from the source tree (this also simplifies the NOTICE
file); licence and copyright notices and links to the online
manpage have been added; mkshrc has been simplified/shrunk;
the shell is always built but not always installed. Together
with ash, /system/bin/sh is now a symlink to either shell.

Switch HAVE_RLIM_T=1 and add to mkmf.sh dummy contents for later…
not tested because this apparently cannot be built on my 32-bit
Kubuntu Hardy any more ☹

Change-Id: I327812d69fff166a1e1f32c1ee3d63f8b48af418