Update patch set 5
Patch Set 5: No score

@Jean-Baptiste: addressed. Now /system/bin/sh is always a
symbolic link to the shell used; both are built but only
one is installed. (The installed one controls the symlink.)

This is so that the EXECUTABLES/{a,mk}sh_intermediates/ are
correctly (re-)used, and not confused between the two shells.

Tested in the emulator with the generic target again, both
“adb shell” and “adb install”. The Dev Tools Terminal also
works. Didn’t test an ash target though.

Patch-set: 5
Label: Code-Review=0