Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I am not 100% certain why I got attached to this issue, but I went through the code in question and have verified that this change should go in. This is actually a rather important change, actually, as right now there is a minor buffer overflow (if more than 1024 bytes is attempted to be sent to the kernel, the original code will set one byte past the end of the buffer to '\0').

Patch-set: 1
Reviewer: Gerrit User 1001184 <1001184@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1