Update patch set 1
Patch Set 1:

Well I saw this work I swear!

What seems to be happening now is watchprops (even an UNMODIFIED watchprops) is always in "R" state, even before ANY prop change is made.  This is because the __futex_wait() call is returning EACCES.  I think it is because the __system_property_area__ is mapped read-only and futex() requires the futex cell be mapped read/write.

If I modify the kernel (where get_futex_key() calls get_user_pages_fast()) to accept read-only cells, it works.

I don't know if you want this patch or not as it is "more" correct than before.  But in addition, either the properties region needs to be mapped read/write or the kernel needs to accept read-only cells.  Modifying the kernel is more correct, IMO, as the futex wait operation doesn't modify the cell as such, but getting the kernel mod accepted may be an uphill battle.  It may be easier to install watchprops setuid and have it map the region read/write.

Patch-set: 1