Update patch set 5
Patch Set 5: Fails

(9 inline comments)

There's quite a variety of style issues that should be improved on - some were explicitly mentioned as comments, but that list isn't exhaustive.

I'm worried about the lack of copyright notices for the new files, as well as the reference to an undefined "original".

Patch-set: 5
Label: Code-Review=0
Label: Verified=-1