Improve ash shell prompt
Make it match mksh (ex: user@host:/cur/dir$), and other shell
distributions, such as bash on Ubuntu.

The host name is essential, because Android-based perhipheral
devices now exist which connect to other master Android
devices. In this case, it's possible to adb shell into the
peripheral from the master device. Without a prompt to tell
which device you're connected to, it's very confusing.

In order to set $HOSTNAME, /etc/profile must be added,
because init.rc is not capable of setting an environment
variable with the output of a call to 'getprop'.

Change-Id: Ia36117dde1f0cd16404afc8d881b7e5df78e1e71