Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

My understanding of SMP on PPC is that it's very similar to ARM, and the sorts of guarantees that you can rely on for x86 do not apply.  If that's the case, this implementation is missing memory barriers in several functions.

See for example http://g.oswego.edu/dl/jmm/cookbook.html -- scroll down to the "Multiprocessors" section and examine the ppc info.

Patch-set: 4
Label: Code-Review=-1
Label: Verified=0