Change GGLSurface.format to integer.
GGLSurface.format is defined as a GGLubyte followed by 3 reserved
GGLubyte values. It appears that this may have been done because
of an Endian issue, but it is clearly incorrect. GGLSurface.format
is indeed an integer (see the value of HAL_PIXEL_FORMAT_YV12 in
hardware/libhardware/include/hardware/hardware.h). Specifically,
there is a test of the format for HAL_PIXEL_FORMAT_YV12 in
frameworks/base/services/surfaceflinger/TextureManager.cpp.

Change-Id: Id07fe271df72c9d327aed617b86197342c8176ec
Signed-off-by: Sean McNeil <sean.mcneil@windriver.com>