init: implement 'capability' option to a service
Helps defining the linux capabilities of a service from init.rc file.
As there is no easy way without file-based capabilities to grant
capabilities beyond an execve, the service is started as root user and
the inheritable capabilities bitmasks is set as specified by
capability option. It is the responsability of the service to setup
requested inheritable capabilities before dropping root.

Signed-off-by: Christian Bejram <christian.bejram@stericsson.com>
Change-Id: I8249140c961cc2296c541a3b228e8fc8eed1c56b