config_mbr: Fix partition table generation
config_mbr() doesn't produce a valid partition table if there was not
one to begin with; the magic sequence 0xAA55 is not written to the last
2 bytes of the MBR. If this MBR signature is missing, the kernel will
refuse to inspect the partition table.

We add a step to config_mbr which blanks out bytes 446-512 of the MBR
and properly writes the MBR signature before doing anything else.
This fixes a problem where diskinstaller had to be run twice if there
was no partition table present on the target volume.

Change-Id: I4a8d173bef231d10e8fbcd0d8aecf54d1cf92474
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>