Add special shutdown and reboot services
Do an actual clean shutdown: kill services and unmount /data first.

Shutdown should never have been done in the framework library. It's
always init's job to bring up and shutdown a system, and init is the
only process with the knowledge of what other processes are running,
and what needs to be killed.

Change-Id: I03e852aaa03931dfd8ee7b553ef9713434cee8ee