Update patch set 1
Patch Set 1: Do not submit

Please don't do that. I already submitted a similar patch a few months ago and it generated much trouble so I reverted it. 

One of the issue is that it changed the behavior of some external libs, for example LLVM would break when the flag is defined (even if it is set to 0) and for upstream compatibility we're not changing the offending code.

If anything, I should probably do a patch commenting in these files that we should not try this again :-)

Patch-set: 1
Reviewer: Gerrit User 1001100 <1001100@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2
Label: Verified=0