Update patch set 1
Patch Set 1:

Thanks for the heads up Raphaël. I'll be glad to refactor this. It appears that your revert is in the internal branch (I don't see it in the external/master branch).

This change is (please note) destined for Gingerbread - so that we can build the IA version (consult with digit on this). Not having this change breaks Dalvik's build -- I can work around that with an #ifndef O_BINARY for just the files that require it. Is that your preference?

Patch-set: 1
Reviewer: Gerrit User 1003360 <1003360@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0