Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

(1 inline comment)

Just a minor nit-pick.

LGTM, but its use is rather limited now when the process is always started as root anyway, and in that case, it can already set capabilities itself before dropping to a less privileged user... Anyway, just my two cents. It seems to work, and that's fine for me

Patch-set: 2