fastboot: set the language local string
Set the language local string to 0x409 (English US / en-us) such that
fastboot will successfully read the usb descriptor serial number string.
The usbdevfs ioctl fails when the wIndex is not initialized with the
local for the unicode of the usb descriptor string.  usbdevfs is an
obsolete interface and fastboot probably needs a re-write to use libusb
like what is in adb.

Signed-off-by: Mark D Horn <mark.d.horn@intel.com>