Make sure data clusters aligned at least 16K border
This ensures the data area start of the sdcard is aligned to at least
16Kb border when formatting.

This was initially for FAT32 that has a cluster size of 32kB.
In the SDA specification the FAT32 file system is offset to align the
data clusters to the physical page size of the NAND Flash used in the
memory cards. If this isn't done an update to a data cluster will
require an update of two physical Flash pages and increase the wear
of the device.

The eMMC devices works in the same way as the memory cards and are
based on very similar NAND Flash, so to limit the wear on the NAND
Flash the logical data clusters should be aligned with the physical
pages. Alignment will also have a positive impact on the performance.

Today's NAND Flash uses a page size of 8 or 16kB.

Change-Id: I45e6d1fc5f45f715f43a8379402bb8830cf86c7e