Modify init.rc and init.goldfish.rc for SE Android.
Set ownerships on /selinux to enable management via Settings app.
If on the emulator, set a policy boolean and restore the context of
/sys/qemu_trace to allow accesses not normally permitted on a device.
Set the security context for the init process.
Restore the security contexts of various runtime directories and files.
Specify the security context for services launched from the rootfs since
we cannot label their executables.

Change-Id: I166ffc267e8e0543732e7118eb0fd4b031efac3b