Update patch set 1
Patch Set 1:

I've noticed that adbd, perhaps the most frequently restarted service, has a good amount of code to workaround this issue for /dev/android_adb and network sockets.

So, I guess the real question is, what _should_ the policy be for service restarts?  In the Linux tradition, we would never expect a new service instance to spawn before the old instance has cleaned up.  It's not clear if simultaneous instances is intentional in Android, or if this is a long standing bug.

Patch-set: 1
Reviewer: Gerrit User 1011168 <1011168@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0