Add a mechanism to build an enviroment partition for a device
This code adds a way that OEMs can create an enviroment partition
for the target device.

This partition can include:
- a fastboot image for the bootloader to display an image during fastboot
- a charger image for the bootloader to display an image during charging
- a splash image the bootloader to display a splash screen
- an enviroment area where persistent data can be stored (i.e lock status,
boot reason or what ever the OEM decided.
- an area for device tree so that OEMs can have a partition for the device tree.

This will allow OEMs to have a single bootloader and a single kernel but a configurable
partition for the device.

Useage:
mkenviroimg --dev_tree <filename> --dt_base <Load address of device tree>
--enviroment <default enviroment data in binary format> --env_size <size>
--board <board name limited to %i characters>
--splash_img <splash image-filename>
--fastboot_img <fastboot image-filename>
--charger_img <charger image-filename>
--fb_base <Framebuffer address for the image>
-o | --output <filename>\n", ENVIRO_NAME_SIZE

This code is abstracted from mkbootimg

Change-Id: Ic8fe44a268e08db61481d02d9f4786cd2d7d6e27
Signed-off-by: Dan Murphy <dmurphy@ti.com>