Update patch set 3
Patch Set 3:

Neither bionic nor the Open/NetBSD projects have the cache that's being protected here.
However all three of them have locking mechanisms in place for the rest of the code, which does not exist in the system/core version. So this commit, if we choose to continue with it, should probably implement the full locking scheme...

I'll get back to you after some more investigation on how I think this should be mitigated.

Patch-set: 3
Label: Code-Review=0