Update patch set 3
Patch Set 3:

Dan, please confirm the Copyright year. If those files are copied from existing files that had a 2007 notice, that's fine, but I'd like to be sure.

Google reviewers, what do you think? This sounds similar to e.g. the /efs partition we've seen on Galaxy Nexus, except that not having it standardized across Android devices means that we have to use proprietary code to access it (fRom, grrrrr).

Patch-set: 3
Label: Code-Review=0