System-core-init: Implement device triggers
Android system/core/init/readme describes two triggers, which are
not implemented in the code yet.
They are:
device-added-<path>
device-removed-<path>
Triggers of these forms occur when a device node is added or removed.
These triggers are very important for PnP devices. For example when
system goes to suspend, the inbuilt usb in some devices is power-off
causing whole driver to be removed. When they are driver restarts on
resume the sysfs files permission will be reset to root again. This
will cause processes with non-root access to fail.
These triggers allows to change ownership again.
Android ueventd.*.rc file allows to modify permissions of file
matching the path sent by udev event, but doesn't allow to set
permissions in sysfs files or other files.
Also this patch goes through device triggers at the end of boot
process and fires them to account for uevents before the init
process was created.

Change-Id: Ib8e6339004ab2e12b2ae0afe7f9b68ebf23df99e
Signed-off-by: srinivas pandruvada <srinivas.pandruvada@intel.com>
Signed-off-by: Daniel Leung <daniel.leung@intel.com>
Reviewed-by: Gumbel, Matthew K <matthew.k.gumbel@intel.com>
Reviewed-by: Jovanovic, Radivoje <radivoje.jovanovic@intel.com>
Reviewed-by: Boie, Andrew P <andrew.p.boie@intel.com>
Tested-by: Boie, Andrew P <andrew.p.boie@intel.com>