System-core-init: Implement device triggers
Android system/core/init/readme describes two triggers, which are
not implemented in the code yet.
They are:
    device-added-<path>
    device-removed-<path>

Triggers of these forms occur when a device node is added or removed.
These triggers are very important for PnP devices. For example when
system goes to suspend, the inbuilt usb in some devices is power-off
causing whole driver to be removed. When they are driver restarts on
resume the sysfs files permission will be reset to root again. This
will cause processes with non-root access to fail.  These triggers
allows to change ownership again.

Android ueventd.*.rc file allows to modify permissions of file matching
the path sent by udev event, but doesn't allow to set permissions in
sysfs files or other files.  Also this patch goes through device triggers
at the end of boot process And fires them to account for uevents before
the init process was created.

Change-Id: Id00c2f3acfd7a5e2e9d132df99a7b9be7acec6cd
Author: srinivas pandruvada <srinivas.pandruvada@intel.com>
Signed-off-by: Xiaokang Qin <xiaokang.qin@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 47435