Update patch set 2
Patch Set 2:

I am going to post a series of patches that adds support to fastboot to use sparse files to flash large files instead of multiflash.  It will use the same getvar:max-download-size variable, so it should be compatible with your bootloader.  The advantage over this patch is that it is transparent to the bootloader as well as to the user - any existing bootloader that supports sparse files will get support for large files.

Patch-set: 2