Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

(me too. this is still broken in internal master, though the code has been lightly refactored so it might not be a trivial merge. i think we're just blocked on needing someone with +2. i volunteer to do the internal merges if no-one else wants to/has time to.)

Patch-set: 2
Label: Code-Review=+1