Update patch set 4
Patch Set 4:

We can change selabel_open in libselinux to take a const argument.

Restart of installd and ueventd appeared to work correctly in our testing.  frameworks appear to allow for installd restart (will re-connect).  However, if this is a problem, we could rework to take an explicit notification of the reload via socket interface.

Patch-set: 4
Label: Code-Review=0