Update patch set 1
Patch Set 1:

The CM team has found devices in the wild that use ro.revision to signal closed-source userspace code how to behave. They also override the hardware revision on the command line, which lead me to this patch.

Perhaps this vendor is doing a bad thing, and this isn't a valid use case. It seemed more like a bug to me, though.

Patch-set: 1