Update patch set 3
Patch Set 3:

Thanks for the review! I fixed the whitespace errors that I had introduced. No functional changes.

Note that there is still trailing whitespace in many of the original files (e.g. the license header of most files), and I did not attempt to fix all these occurrences. If you like, I'd be happy to do that after this CL is reviewed.

Patch-set: 3
Label: Code-Review=0