Update patch set 5
Patch Set 5:

To clarify, /selinux/booleans is initially populated by the initial policy load from the booleans defined in the policy, and re-populated upon a policy reload.  Since all policy loading is done via init in SE Android, it doesn't need to bother with the netlink notifications for this purpose.  So we could hardcode logic in init to walk /selinux/booleans after policy load and set ownerships to system UID if that would be preferable.

Patch-set: 5