Update patch set 8
Patch Set 8:

We are ready to push an updated patch set directly related to Dima's comments.  A side effect of the changes is that a dependency now exists with  If6af6c4887cdead949737cebdd673957e9273ead. That is, the new code needs changes made by the listed change-id. Any chance we can get this merged soon so that we can avoid yet another dependency with our code?

Patch-set: 8