Update patch set 1
Patch Set 1:

> Extend ueventd.rc by supporting one more optional column.
That is what I meant by """<normal rule> trigger:my_id""" .

The reason for the keyword "trigger:", is that it might make parsing easier, and is more readable. But that's just an idea.

> This column can be used to set some property.
> This way it can fire some init script property using "On property"..

That is what I meant by """passed to the normal init triggering mechanism "<name>=<value>" """ and by """ which would cause a "setprop uevent_my_id={add,remove,...}" """

> on property:snd._device_present=1
Bad idea.
You won't be able to distinguish added vs removed. Which is what the original design called for. So go with what I was saying. Which would look like:

  /dev/snd/* 0660 system system trigger:snd_device

  on property:uevent_snd_device=add
    do_something_related_to_add
  on property:uevent_snd_device=remove
    do_something_related_to_remove

Patch-set: 1