Update patch set 3
Patch Set 3:

I agree with Dima that it's confusing that the limits are applied in reverse order, however when I implemented a version that added elements to the end of the list the code just looked messy. This combined with adding elements to the front of the list is what is done in the other cases; e.g sockets.

One way to solve this would be to use the helper functions from <sys/queue.h>. Let me know what you think and I can give it a try.

Updated in this version is naming of the function for resource-lookup, an function for reverse lookup is added and this is used to print reasonable error messages.

Also added is the use of the name lookup to the setrlimits method, to give consistency between the global and service specific ways of setting rlimits. Both methods accept both numbers and keywords.

Patch-set: 3
Reviewer: Gerrit User 1011077 <1011077@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0