Update patch set 3
Patch Set 3:

Note that we do not have much time left to review this patch to make it included for the next release. Bigger patch will greatly increase the risk of finishing the review in time. And I am not sure if generic C version works without any change. If that is not the case, we may need a patch to make generic C version work first as a backup plan.

Patch-set: 3