Update patch set 4
Patch Set 4:

This changeset removes the MIPS big-endian support, and does some other small cleanup. This should make the code easier to review.

While the generic pixel pipeline can work for MIPS, it is of huge benefit to us to have this much faster rendering support in the emulator. I hope that you can find the time to review this, and I will respond quickly with updates as needed.

I will add some explanation here about the structure of the code, and why I believe that this will not impact the existing ARM port. Please also see the more concise explanation at the top of MIPSAssembler.cpp.

While adding the MIPS support to to the existing Pixelflinger, there are 3 areas of the existing code that are altered.

1. The ARMAssemblerInterface appears to be designed to be subclassed to multiple assemblers, and MIPS takes advantage of that. However there are 14 functions to set up ARM addressing modes that were defined as static functions. These have been changed to virtual, so they can also be overridden in the MIPS assembler, as we do with all the other assembler methods (which were already virtual). I moved the implementation from ARMAssemblerInterface to ARMAssembler, without changing the functionality. Since these functions are so central to the working of the ARM assembler, defects here would NOT be subtle. We would see catastrophic failure, and that is not the case.

With a couple very small exceptions, ALL changes in the ARMAssembler* files are for this simple purpose.

2. The register allocator in GGLAssembler.cpp has a conceptually trivial change to map ARM registers 0-15 to MIPS registers 2-17, by adding 2 to the value in a few places. This value (mRegisterOffset) is initialized in the class constructor to 0, and ONLY changed to value 2 for MIPS. As above, a defect in this area would not have subtle effects. The ARM version would fail hard and fast if this were broken, and it does not.

3. The original GGLAssembler was written to try to generate optimal ARM code, and if the the ARM register pool was exhausted, it would throw the code away, and retry at a lower optimization level. There are several places in the original GGLAssembler where the register allocator status is checked, and the code bails out for a retry. During testing, I found a few cases where we needed to bail out like this in the texturing module (which does generate a lot of register pressure). I have added some registerFile status checks and bailouts there. These should help ARM as well as MIPS in some rare cases.

Finally, long ago, I had removed several instances of trailing whitespace in  GGLAssembler.cpp, which now shows up in the diff. I can restore these if you want, but I hope you can just overlook the whitespace changes.

Thanks for your time in reviewing this.

Patch-set: 4
Label: Code-Review=0