Add exec support by flag
This patch will add exec support to init by the flag "NEEDS_EXEC_SUPPORT".
It would be good to have this on some devices like the Sony LT26.
Example for the usage would be the boot of recovery from bootpartition, because
this device lacks a standard-recovery partition. On this device recovery usually
gets booted in 3rd party roms by executing a shellscript to grab the keyinputs
and then boot recovery on a predefined keycombination.
Also it would be good to have this included for chargemon on this device.

Change-Id: Iec16aed306e796989a7506faf956473714e308f9
Signed-off-by: Andreas Makris <andreas.makris@gmail.com>