Update patch set 1
Patch Set 1:

It appears that this commit assumes video hardware that has its own framebuffer. If the bootloader framebuffer resides in memory that linux will own after boot (which is the case on all our platforms), then that framebuffer will likely be clobbered during linux boot.

It seems to me that the solution for the scenario described in the commit message should be the absence of INIT_IMAGE_FILE in the filesystem. If you don't want it shown, don't include it, but this commit would break platforms that need it.

Patch-set: 1
Reviewer: Gerrit User 1015159 <1015159@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0