property: Fix Encryption design issue
Currently, Encryption feature works as below:
1. stop all services in main class, including the ones disabled by default
   and triggered by properties, such as aplogd;
2. load persist props from /data/property to care user changed properties;
3. restart services in main class, pay attention here, only non-disabled
   services would be started here.

So, the issue is mainly on step2.
We only read /data/property, and missed the ones that users never changed,
such as persist.log.aplogd.enable. So that, the services depending on those
properties could not be really restarted.

To resove the issue:
we would re-load all default properties from /system/build.prop and
/system/default.prop, and then load the user changed properties
from /data/property/.

Change-Id: Ia94ead9d525d39fabe193ddd8acc28e2ae57c41d
Signed-off-by: Hong-Mei Li <a21834@motorola.com>