Update patch set 2
Patch Set 2:

The patch was to fix ifconfig output on bigendian targets (the original code reads the (bigendian) network address as a long and then unpacks it assuming the MS byte is in bits 7..0), but this seems like a good change to have anyway. The resulting code is simpler and uses a standard libc function to format the IP address.

Patch-set: 2
Reviewer: Gerrit User 1003284 <1003284@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0