RFC: init: optionally create by-name links for GPT partitions
--------------------------------------------------------------
[Submitter's note] Requesting comments on this patch. Currently
there is already something like this in ueventd, but the name
of the disk controller becomes part of the generated symlink
path. By just having /dev/block/by-name/xxxxxx we never have
to change any of our fstab files, and we are able to do things
like dual boot with other OSes or have a single set of fstab
files for a variety of boards that have very different hardware.
--------------------------------------------------------------

Create device node symlinks by name if the uevent for a block
device has a name in it, and the name passed in matches the
prefix for the installation ID. This prevents collisions with
multiple Android installations on the same system. The created
symlinks do not have the prefix in them.

If the installation ID isn't set (through ro.boot.install_id)
this mechanism doesn't do anything.

This mechanism, when used, allows for the use of stable by-name
symlinks to device nodes in the various disk configuration files
(mountall fstab, recovery.fstab, vold.conf, etc), with the actual
mapping to physical disk partitions done in the GPT.

Change-Id: Icee3e86bef5569c2bbd94c26bc00d49028345e3b
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>