fs_mgr: Make sure to always check the fs that actually is mounted
Changing so check_fs() is done on the provided loopback file system
instead of the hardware device specified in the fstab file. These can
be the same, but for example in case of encrypted file systems they
are not. The provided loopback file system is the one that is mounted,
so that is the one that always should be checked.

Change-Id: Id41e2e45572970e184cdb1d2eadc064e652de230