Update patch set 1
Patch Set 1:

This change is required to allow adb to be used from Python on Windows. It prevents the 'forked' adb server process from inheriting the stdout and stderr handles, otherwise, the calling program (in this case Python) gets stuck reading from the read handle of these pipes as the 'forked' adb server retains a reference to the write handle.

Patch-set: 1