init: util: create hardlinks for device nodes
init has some logic where it creates symlinks to device nodes under
/dev, for example for block device nodes that are named according to
their label. It's impossible to change the permissions or ownership
on these device nodes in init.rc if referred to by their symlink, for
security reasons symbolic links are not followed in chown/chmod.
Change ueventd to create hardlinks to the device nodes instead.

Change-Id: I1c2930cfdb635762a7263c452580c720bd4873f7
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>